Add files via upload
Here I designed an Expressway Tool Both Systems, where Road has 2 lanes, one lane for the entry expressway and another is to out. Day Night mode also add here. when the sun is out from the window then the view will be night and to go back to day mode we have to press 'd' or 'D'. When we press 'a' or 'A' the traffic police will hand up and the signal light will be red and pressing 's' or 'S' we can stop car and when press to 'z' the traffic police hand down and signal light will be green and the car will start moving again. Here I also use some more keyboard function like pressing 'q' will stop the bus and pressing 'z' will move it again.